+++ example/migrations/1567307130689-using-env-variable-defined-in-serverless.js
@@ -4,6 +4,7 @@ module.exports.up = function (next) {
   console.log(`do action using DATABASE_NAME=${process.env.DATABASE_NAME}`);
   console.log(`do action using ANOTHER_ENV=${process.env.ANOTHER_ENV}`);
   console.log(`The root folder is ${process.env.SERVERLESS_ROOT_PATH}`);
+  console.log(`The complex var is ${process.env.COMPLEX_VAR}`);
   next();
 };
 
@@ -11,6 +12,7 @@ module.exports.down = function (next) {
   console.log(`undo action using DATABASE_NAME=${process.env.DATABASE_NAME}`);
   console.log(`undo action using ANOTHER_ENV=${process.env.ANOTHER_ENV}`);
   console.log(`The root folder is ${process.env.SERVERLESS_ROOT_PATH}`);
+  console.log(`The complex var is ${process.env.COMPLEX_VAR}`);
   next();
 };