+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### Before release 2.0.0 - 2020-02-07
+## Before release 2.0.2 - 2020-02-10
+### Fixed
+- Serverless config variables are now interpolated into env variables.
+## Before release 2.0.1 - 2020-02-07
+### Fixed
+- When `store` or `state-file` are not specified it will not fail.
+## Before release 2.0.0 - 2020-02-07
 ### Added
 - Allows to override environment variables
 ### Changed